detection-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic-settings==2.6.0
anthropic>=0.39.0
pillow>=10.0.0
pillow-heif>=0.16.0
